Refactor XVLA model and checkpoint save modules
Rename the XVLA training wrapper to X_VLA, fold transformer blocks into the XVLA flow-matching head, and update Florence2/action-space integrations.

Remove trainable_module_keys dependencies from VLA modules so runner checkpointing can save full model modules.
