+++ fluxvla/models/vlas/__init__.py
@@ -19,7 +19,7 @@ from .pi05_flowmatching import PI05FlowMatching  # noqa: F401, F403
 from .pi05_flowmatching_inference import \
     PI05FlowMatchingInference  # noqa: F401, F403
 from .smolvla_flowmatching import SmolVLAFlowMatching  # noqa: F401, F403
-from .xvla_flowmatching import XVLAFlowMatching  # noqa: F401, F403
+from .x_vla import X_VLA  # noqa: F401, F403
 
 try:
     from .dreamzero_vla import DreamZeroVLA  # noqa: F401
